Write the find_array_function (details in description)
The find_array_function loops through the char array being disemvoweled and checks every character by calling the detect_vowel function. For characters whose values are not vowels, the function increments a counter that is returned to the caller. This counter begins at 1 so that an extra space is left for a null terminator in an array of length find_array_length.