Write the disemvowel function and add comments
The disemvowel function creates an array of length find_array_size(str) --an array sufficiently long for the disemvoweled output with space for a null terminator. After that, the function loops through the input array str and only copies non-vowel characters into the new array (with help of the detect_vowel function).